refactor(rs-dapi-client): move AddressBanInfo to its own module
Split the public AddressBanInfo snapshot out of address_list.rs into
address_ban_info.rs (address_list.rs now holds only Address /
AddressStatus / AddressList). The crate-root re-export path
(rs_dapi_client::AddressBanInfo) is unchanged, so rs-sdk / platform-wallet
consumers are unaffected. Drive-by: fixed a pre-existing broken intra-doc
link ([get_live_address] -> [`Self::get_live_address`]) in the same file.
